How i solved the diagonal difference problem
This problem is under the easy category, but the joy of solving it is fulfilling. 
We are given a list of numbers representing the number of rows + a square matrix... we are supposed to find the positive difference between the addition of two longest diagonals.  This gives me so much joy because i solved myself... and the solution isn't even anywhere online and i got the answer.. it works for all test cases... it may not be effective but at this point i am barely two months into learning python